Enhance enrollment services and improve unit test coverage

+++ backend/src/main/java/com/backend/dto/EnrollmentDTO.java
@@ -5,12 +5,12 @@ import java.util.Date;
 
 @Data
 public class EnrollmentDTO {
-    private Long userId;
+    private Integer userId;
     private Long courseId;
     private Date enrollmentDate;
     private String grade;
 
-    public void setUserId(Long userId) {
+    public void setUserId(Integer userId) {
         this.userId = userId;
     }
     public void setCourseId(Long courseId) {
@@ -22,7 +22,7 @@ public class EnrollmentDTO {
     public void setGrade(String grade) {
         this.grade = grade;
     }
-    public Long getUserId() {
+    public Integer getUserId() {
         return userId;
     }
     public Long getCourseId() {
